Fix problem finding local dependencies of dynamic libraries in Windows
If the library loaded is not in the program's library path, any of the
library's own dependencies in the same directory as the library won't be
found normally.  To overcome this, temporarily change the current
directory to the directory containing the library, and change back after
the library is loaded.

Note that changing the current directory applies to the entire process,
and so is not thread-safe.