Fix File::Clone bug for Windows
The Windows implementation was duplicating the file handle,
without guaranteeing that both handles ended up with
file pointers that were independent of each other.

To fix this bug, use CreateFile to make the clone, as is
recommended in the MSDN page for DuplicateHandle, in the
case where independent file pointers is desired.