DataQueue: fix nasty bug in dequeue
No idea why this went unnoticed for so long (4 months?)...

This is a nasty bug where the queue has enough data but
if it's fragmented you won't be able to peek or remove
it all in one shot if the fragments are certain sizes.

The local variable used to track the number of bytes remaining
in the queue was being subtracted by the wrong value.